+++ lua/obsidian/commands/new_from_template.lua
@@ -4,7 +4,7 @@ local api = require "obsidian.api"
 local Note = require "obsidian.note"
 
 ---@param data CommandArgs
-return function(_, data)
+return function(data)
   local picker = Obsidian.picker
   if not picker then
     log.err "No picker configured"