+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - remove custom enumerate function to use vim.iter.
 - `blink` is checked before `nvim_cmp`.
 - `opts.note_frontmatter_func` defaults to `require"builtin".frontmatter`.
+- `Client.find_tags` -> `search.find_tags`.
+- all commands and callbacks now don't have client as callback.
 
 ### Fixed